Improving solver.
Working on a more sophisiticated solver for sum puzzles.
Now have a way to create a reduced works grid by taking the sums into account.
It can now determine whether a sum constraint can be satisifed by
examining the works grid.  Any cells in the sum area that have only one
possibility are removed from the area and the required sum.  If only one
cell remains, then it sees if that cell can have the required value. This
allows certain cases to be quickly rejected that weren't rejected before.